Move MobileBookingHeader to a separate file
